requirements.txt
gspread
oauth2client
qrcode
pillow
pandas
pyfiglet
regex
unittest2
pyzbar